+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - (Map Editor) Fixed crash when fusing vertices
+- (Sprite Editor) Copy animation works properly